+++ packages/gatsby-theme-aoi/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.9.0](https://github.com/cieloazul310/gatsby-aoi/compare/v3.0.0-alpha.6...v2.9.0) (2022-11-11)
+
+**Note:** Version bump only for package @cieloazul310/gatsby-theme-aoi
+
+# [2.8.0](https://github.com/cieloazul310/gatsby-aoi/compare/v3.0.0-alpha.6...v2.8.0) (2022-11-11)
+
+**Note:** Version bump only for package @cieloazul310/gatsby-theme-aoi
+
 # [3.0.0-alpha.6](https://github.com/cieloazul310/gatsby-aoi/compare/v3.0.0-alpha.5...v3.0.0-alpha.6) (2022-11-08)
 
 **Note:** Version bump only for package @cieloazul310/gatsby-theme-aoi